不恰好だが、日付が矢印で動くようになった
