+++ src/App.tsx
@@ -1,5 +1,5 @@
 import CalendarCalculator from './CalendarElement';
-import Navigation from './Navigation';
+import { Navigation } from './Navigation';
 
 const App = () => {
   return (
